// backend/types/figmaApiTypes.ts
export type Node =
	| BooleanOperationNode
	| ComponentNode
	| ComponentSetNode
	| ConnectorNode
	| EllipseNode
	| EmbedNode
	| FrameNode
	| GroupNode
	| InstanceNode
	| LineNode
	| LinkUnfurlNode
	| RectangleNode
	| RegularPolygonNode
	| SectionNode
	| ShapeWithTextNode
	| SliceNode
	| StarNode
	| StickyNode
	| TableNode
	| TableCellNode
	| TextNode
	| VectorNode
	| WashiTapeNode
	| WidgetNode
	| DocumentNode
	| CanvasNode;

export type DocumentNode = {
	type: "DOCUMENT";

	children: CanvasNode[];
} & IsLayerTrait;

export type CanvasNode = {
	type: "CANVAS";

	children: SubcanvasNode[];

	/**
	 * Background color of the canvas.
	 */
	backgroundColor: RGBA;

	/**
	 * Node ID that corresponds to the start frame for prototypes. This is deprecated with the
	 * introduction of multiple flows. Please use the `flowStartingPoints` field.
	 *
	 * @deprecated
	 */
	prototypeStartNodeID: string | null;

	/**
	 * An array of flow starting points sorted by its position in the prototype settings panel.
	 */
	flowStartingPoints: FlowStartingPoint[];

	/**
	 * The device used to view a prototype.
	 */
	prototypeDevice: PrototypeDevice;

	measurements?: Measurement[];
} & IsLayerTrait &
	HasExportSettingsTrait;
// ...
export type BooleanOperationNode = {
	/**
	 * The type of this node, represented by the string literal "BOOLEAN_OPERATION"
	 */
	type: "BOOLEAN_OPERATION";

	/**
	 * A string enum indicating the type of boolean operation applied.
	 */
	booleanOperation: "UNION" | "INTERSECT" | "SUBTRACT" | "EXCLUDE";
} & IsLayerTrait &
	HasBlendModeAndOpacityTrait &
	HasChildrenTrait &
	HasLayoutTrait &
	HasGeometryTrait &
	HasExportSettingsTrait &
	HasEffectsTrait &
	HasMaskTrait &
	TransitionSourceTrait;

export type SectionNode = {
	/**
	 * The type of this node, represented by the string literal "SECTION"
	 */
	type: "SECTION";

	/**
	 * Whether the contents of the section are visible
	 */
	sectionContentsHidden: boolean;
} & IsLayerTrait &
	HasGeometryTrait &
	HasChildrenTrait &
	HasLayoutTrait &
	DevStatusTrait;

export type FrameNode = {
	/**
	 * The type of this node, represented by the string literal "FRAME"
	 */
	type: "FRAME";
} & FrameTraits;

export type GroupNode = {
	/**
	 * The type of this node, represented by the string literal "GROUP"
	 */
	type: "GROUP";
} & FrameTraits;

export type ComponentNode = {
	/**
	 * The type of this node, represented by the string literal "COMPONENT"
	 */
	type: "COMPONENT";
} & FrameTraits &
	ComponentPropertiesTrait;

export type ComponentSetNode = {
	/**
	 * The type of this node, represented by the string literal "COMPONENT_SET"
	 */
	type: "COMPONENT_SET";
} & FrameTraits &
	ComponentPropertiesTrait;

export type VectorNode = {
	/**
	 * The type of this node, represented by the string literal "VECTOR"
	 */
	type: "VECTOR";
} & CornerRadiusShapeTraits &
	AnnotationsTrait;

export type StarNode = {
	/**
	 * The type of this node, represented by the string literal "STAR"
	 */
	type: "STAR";
} & CornerRadiusShapeTraits &
	AnnotationsTrait;

export type LineNode = {
	/**
	 * The type of this node, represented by the string literal "LINE"
	 */
	type: "LINE";
} & DefaultShapeTraits &
	AnnotationsTrait;

export type EllipseNode = {
	/**
	 * The type of this node, represented by the string literal "ELLIPSE"
	 */
	type: "ELLIPSE";

	arcData: ArcData;
} & DefaultShapeTraits &
	AnnotationsTrait;

export type RegularPolygonNode = {
	/**
	 * The type of this node, represented by the string literal "REGULAR_POLYGON"
	 */
	type: "REGULAR_POLYGON";
} & CornerRadiusShapeTraits &
	AnnotationsTrait;

export type RectangleNode = {
	/**
	 * The type of this node, represented by the string literal "RECTANGLE"
	 */
	type: "RECTANGLE";
} & RectangularShapeTraits;

export type TextNode = {
	/**
	 * The type of this node, represented by the string literal "TEXT"
	 */
	type: "TEXT";
} & DefaultShapeTraits &
	TypePropertiesTrait &
	AnnotationsTrait;

export type TableNode = {
	/**
	 * The type of this node, represented by the string literal "TABLE"
	 */
	type: "TABLE";
} & IsLayerTrait &
	HasChildrenTrait &
	HasLayoutTrait &
	MinimalStrokesTrait &
	HasEffectsTrait &
	HasBlendModeAndOpacityTrait &
	HasExportSettingsTrait;

export type TableCellNode = {
	/**
	 * The type of this node, represented by the string literal "TABLE_CELL"
	 */
	type: "TABLE_CELL";
} & IsLayerTrait &
	MinimalFillsTrait &
	HasLayoutTrait &
	HasTextSublayerTrait;

export type SliceNode = {
	/**
	 * The type of this node, represented by the string literal "SLICE"
	 */
	type: "SLICE";
} & IsLayerTrait;

export type InstanceNode = {
	/**
	 * The type of this node, represented by the string literal "INSTANCE"
	 */
	type: "INSTANCE";

	/**
	 * ID of component that this instance came from.
	 */
	componentId: string;

	/**
	 * If true, this node has been marked as exposed to its containing component or component set.
	 */
	isExposedInstance?: boolean;

	/**
	 * IDs of instances that have been exposed to this node's level.
	 */
	exposedInstances?: string[];

	/**
	 * A mapping of name to `ComponentProperty` for all component properties on this instance. Each
	 * property has a type, value, and other optional values.
	 */
	componentProperties?: { [key: string]: ComponentProperty };

	/**
	 * An array of all of the fields directly overridden on this instance. Inherited overrides are not
	 * included.
	 */
	overrides: Overrides[];
} & FrameTraits;

export type EmbedNode = {
	/**
	 * The type of this node, represented by the string literal "EMBED"
	 */
	type: "EMBED";
} & IsLayerTrait &
	HasExportSettingsTrait;

export type LinkUnfurlNode = {
	/**
	 * The type of this node, represented by the string literal "LINK_UNFURL"
	 */
	type: "LINK_UNFURL";
} & IsLayerTrait &
	HasExportSettingsTrait;

export type StickyNode = {
	/**
	 * The type of this node, represented by the string literal "STICKY"
	 */
	type: "STICKY";

	/**
	 * If true, author name is visible.
	 */
	authorVisible?: boolean;
} & IsLayerTrait &
	HasLayoutTrait &
	HasBlendModeAndOpacityTrait &
	MinimalFillsTrait &
	HasMaskTrait &
	HasEffectsTrait &
	HasExportSettingsTrait &
	HasTextSublayerTrait;

export type ShapeWithTextNode = {
	/**
	 * The type of this node, represented by the string literal "SHAPE_WITH_TEXT"
	 */
	type: "SHAPE_WITH_TEXT";

	/**
	 * Geometric shape type. Most shape types have the same name as their tooltip but there are a few
	 * exceptions. ENG_DATABASE: Cylinder, ENG_QUEUE: Horizontal cylinder, ENG_FILE: File, ENG_FOLDER:
	 * Folder.
	 */
	shapeType: ShapeType;
} & IsLayerTrait &
	HasLayoutTrait &
	HasBlendModeAndOpacityTrait &
	MinimalFillsTrait &
	HasMaskTrait &
	HasEffectsTrait &
	HasExportSettingsTrait &
	HasTextSublayerTrait &
	CornerTrait &
	MinimalStrokesTrait;

export type ConnectorNode = {
	/**
	 * The type of this node, represented by the string literal "CONNECTOR"
	 */
	type: "CONNECTOR";

	/**
	 * The starting point of the connector.
	 */
	connectorStart: ConnectorEndpoint;

	/**
	 * The ending point of the connector.
	 */
	connectorEnd: ConnectorEndpoint;

	/**
	 * A string enum describing the end cap of the start of the connector.
	 */
	connectorStartStrokeCap:
		| "NONE"
		| "LINE_ARROW"
		| "TRIANGLE_ARROW"
		| "DIAMOND_FILLED"
		| "CIRCLE_FILLED"
		| "TRIANGLE_FILLED";

	/**
	 * A string enum describing the end cap of the end of the connector.
	 */
	connectorEndStrokeCap:
		| "NONE"
		| "LINE_ARROW"
		| "TRIANGLE_ARROW"
		| "DIAMOND_FILLED"
		| "CIRCLE_FILLED"
		| "TRIANGLE_FILLED";

	/**
	 * Connector line type.
	 */
	connectorLineType: ConnectorLineType;

	/**
	 * Connector text background.
	 */
	textBackground?: ConnectorTextBackground;
} & IsLayerTrait &
	HasLayoutTrait &
	HasBlendModeAndOpacityTrait &
	HasEffectsTrait &
	HasExportSettingsTrait &
	HasTextSublayerTrait &
	MinimalStrokesTrait;

export type WashiTapeNode = {
	/**
	 * The type of this node, represented by the string literal "WASHI_TAPE"
	 */
	type: "WASHI_TAPE";
} & DefaultShapeTraits;

export type WidgetNode = {
	/**
	 * The type of this node, represented by the string literal "WIDGET"
	 */
	type: "WIDGET";
} & IsLayerTrait &
	HasExportSettingsTrait &
	HasChildrenTrait;
// ...
export type AltNode = Node & {
	styledTextSegments: Array<Pick<StyledTextSegment, any | "characters" | "start" | "end">>;
	cumulativeRotation: number;
	uniqueName: string;
	canBeFlattened: boolean;
	isRelative: boolean;
	width: number;
	height: number;
	x: number;
	y: number;
};